% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFok.R
\name{plot.rfok}
\alias{plot.rfok}
\title{Plot method for rfok}
\usage{
\method{plot}{rfok}(x, conf = 0.95, ...)
}
\arguments{
\item{x}{Object returned by \code{rfok()}}

\item{conf}{Confidence level for plotting}

\item{...}{Optional parameters to pass to \code{plot()}}
}
\value{
An object with class "rfok"
}
\description{
A conformal RF based on Johansson et al. (2014). Rather than using a traditional calibration set, conformal intervals are constructed using out-of-bag samples for each tree. K-nearest neighbors is used to calibrate the prediction intervals for each prediction location.
}
\examples{
# Friedman function
f <-  function(x){
  10 * sin(pi * x[1] * x[2]) + 20 * (x[3] - 0.5)^2 + 10 * x[4] + 5 * x[5]
}
X <- matrix(runif(250), nrow=50, ncol=5)
y <- apply(X, 1, f)
fit <- rfok(X, y)

plot(fit)
}
